+++ store/repos/type.go
@@ -52,6 +52,7 @@ type nullBuild struct {
 	Params       types.JSONText
 	Cron         sql.NullString
 	Deploy       sql.NullString
+	DeployID     sql.NullInt64
 	Started      sql.NullInt64
 	Finished     sql.NullInt64
 	Created      sql.NullInt64
@@ -91,6 +92,7 @@ func (b *nullBuild) value() *core.Build {
 		Params:       params,
 		Cron:         b.Cron.String,
 		Deploy:       b.Deploy.String,
+		DeployID:     b.DeployID.Int64,
 		Started:      b.Started.Int64,
 		Finished:     b.Finished.Int64,
 		Created:      b.Created.Int64,