+++ store/shared/migrate/sqlite/files/013_add_column_builds_deploy_id.sql
@@ -1,0 +1,3 @@
+-- name: alter-table-builds-add-column-deploy-id
+
+ALTER TABLE builds ADD COLUMN build_deploy_id NUMBER NOT NULL DEFAULT 0;