+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Support for basic auth as an option for Gitea, by [@techknowlogick](https://giteahub.com/techknowlogick). [#2721](https://github.com/drone/drone/issues/2721)
 
+## [1.3.1] - 2019-08-26
+### Added
+- support for the GitHub deployment status API, by [@bradrydzewski](https://github.com/bradrydzewski).
+
 ## [1.3.0] - 2019-08-20
 ### Added
 - support for storing logs in Azure Cloud Storage, by [@Lucretius](https://github.com/Lucretius). [#2788](https://github.com/drone/drone/pull/2788)